requirements.txt
flask
pandas
requests
openpyxl
lxml
python-dotenv
html5lib
beautifulsoup4